cleanup codes for SQL playground
