+++ packages/legend-application-studio/src/stores/editor/ExplorerTreeState.ts
@@ -51,7 +51,7 @@ import {
   type RelationalDatabaseConnection,
 } from '@finos/legend-graph';
 import { APPLICATION_EVENT } from '@finos/legend-application';
-import { DatabaseBuilderState } from './editor-state/element-editor-state/connection/DatabaseBuilderState.js';
+import { DatabaseBuilderWizardState } from './editor-state/element-editor-state/connection/DatabaseBuilderWizardState.js';
 
 export enum ExplorerTreeRootPackageLabel {
   FILE_GENERATION = 'generated-files',
@@ -75,7 +75,7 @@ export class ExplorerTreeState {
 
   elementToRename?: PackageableElement | undefined;
   classToGenerateSampleData?: Class | undefined;
-  databaseBuilderState: DatabaseBuilderState | undefined;
+  databaseBuilderState: DatabaseBuilderWizardState | undefined;
 
   constructor(editorStore: EditorStore) {
     makeObservable(this, {
@@ -103,7 +103,7 @@ export class ExplorerTreeState {
       buildTreeInTextMode: action,
       openExplorerTreeNodes: action,
       reprocess: action,
-      buildDbBuilderState: action,
+      buildDatabase: action,
       setDatabaseBuilderState: action,
       onTreeNodeSelect: action,
       openNode: action,
@@ -176,14 +176,11 @@ export class ExplorerTreeState {
     this.classToGenerateSampleData = val;
   }
 
-  setDatabaseBuilderState(val: DatabaseBuilderState | undefined): void {
+  setDatabaseBuilderState(val: DatabaseBuilderWizardState | undefined): void {
     this.databaseBuilderState = val;
   }
-  buildDbBuilderState(
-    val: RelationalDatabaseConnection,
-    isReadOnly: boolean,
-  ): void {
-    const dbBuilderState = new DatabaseBuilderState(
+  buildDatabase(val: RelationalDatabaseConnection, isReadOnly: boolean): void {
+    const dbBuilderState = new DatabaseBuilderWizardState(
       this.editorStore,
       val,
       isReadOnly,
