+++ packages/legend-application-studio/style/components/editor/_database-builder.scss
@@ -128,7 +128,7 @@
   }
 }
 
-.database-builder-tree {
+.database-schema-explorer {
   height: 100%;
   width: 100%;
   display: flex;
